Change layout of ErrorMessages
